packages/materials/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "firecore-materials"
version = "0.1.0"
description = "FireCore prefab material catalog — SIPs, CLT, steel with climate ratings"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["firecore_materials*"]
